Remove core dependencies from generic part
This concerns:
- Communication Non-Secure <--> Secure
- sys/types.h contains some types not defined on all compilers
