+++ core/arch/arm32/tee/tee_rpmb.c
@@ -24,7 +24,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-#include <sys/types.h>
+#include <types_ext.h>
 #include <stdlib.h>
 #include <string.h>
 #include <string_ext.h>
