+++ lib/libutils/ext/strlcat.c
@@ -34,7 +34,7 @@ static char *rcsid =
 	"$OpenBSD: strlcat.c,v 1.8 2001/05/13 15:40:15 deraadt Exp $";
 #endif /* LIBC_SCCS and not lint */
 
-#include <sys/types.h>
+#include <types_ext.h>
 #include <string_ext.h>
 /* #include <string.h> */
