+++ lib/libutils/ext/strlcpy.c
@@ -34,7 +34,7 @@ static char *rcsid =
 	"$OpenBSD: strlcpy.c,v 1.5 2001/05/13 15:40:16 deraadt Exp $";
 #endif /* LIBC_SCCS and not lint */
 
-#include <sys/types.h>
+#include <types_ext.h>
 #include <string_ext.h>
 
 /*
